+++ app/src/main/java/com/olshevchenko/currencyconverter/datasource/RatesDataSource.kt
@@ -1,9 +1,0 @@
-package com.olshevchenko.currencyconverter.datasource
-
-import com.olshevchenko.currencyconverter.core.Result
-import com.olshevchenko.currencyconverter.core.data.entity.RatesDataEntity
-import io.reactivex.Single
-
-interface RatesDataSource {
-    fun getRates(): Single<Result<RatesDataEntity>>
-}