+++ app/src/main/java/com/olshevchenko/currencyconverter/features/rates/domain/usecase/GetCurrencyRatesUseCase.kt
@@ -2,7 +2,6 @@ package com.olshevchenko.currencyconverter.features.rates.domain.usecase
 
 import com.olshevchenko.currencyconverter.features.rates.domain.repository.CurrencyRatesRepository
 import com.olshevchenko.currencyconverter.core.Result
-import com.olshevchenko.currencyconverter.core.data.entity.RatesDataEntity
 import com.olshevchenko.currencyconverter.core.usecase.RXUseCase
 import com.olshevchenko.currencyconverter.features.rates.domain.model.CurrencyRates
 import io.reactivex.Scheduler
@@ -11,9 +10,9 @@ import io.reactivex.Single
 class GetCurrencyRatesUseCase(
     private val ratesRepository: CurrencyRatesRepository,
     subscribeScheduler: Scheduler,
-    private val postExecutionScheduler: Scheduler
-) :
-    RXUseCase<Result<CurrencyRates>, Unit>(subscribeScheduler, postExecutionScheduler) {
+    observeScheduler: Scheduler
+) : RXUseCase<Result<CurrencyRates>, Unit>(subscribeScheduler, observeScheduler) {
+
     override fun buildUseCaseSingle(params: Unit?): Single<Result<CurrencyRates>> =
         ratesRepository.getRates()
 }