+++ app/src/main/java/com/olshevchenko/currencyconverter/features/rates/presentation/RatesListView.kt
@@ -1,0 +1,14 @@
+package com.olshevchenko.currencyconverter.features.rates.presentation
+
+class RatesListView : RatesContract.View {
+
+    private var ratesPresenter: RatesContract.Presenter? = null
+
+    override fun attachPresenter(presenter: RatesContract.Presenter) {
+        ratesPresenter = presenter
+    }
+
+    override fun initView() {}
+
+    override fun setViewState(viewState: RatesViewState) {}
+}